% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_comext.R
\name{add_quarter_column}
\alias{add_quarter_column}
\title{Convert monthly data to quarterly periods}
\usage{
add_quarter_column(data)
}
\description{
Convert monthly data to quarterly periods
}
\keyword{internal}
