% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_comext.R
\name{check_quarter_completeness}
\alias{check_quarter_completeness}
\title{Check if all months are available for complete quarters}
\usage{
check_quarter_completeness(data, target_quarter)
}
\description{
Check if all months are available for complete quarters
}
\keyword{internal}
