% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_a_plots_T_V.R
\name{nfsa_a_plots_T_V}
\alias{nfsa_a_plots_T_V}
\title{Generate plots comparing not adjusted data from two sources (T and V) for a given country.}
\usage{
nfsa_a_plots_T_V(
  country,
  output_sel = here::here("output", "plots"),
  time_min = 1995,
  my_theme = ggthemes::theme_fivethirtyeight(),
  my_colours = c("#B656BD", "#208486")
)
}
\arguments{
\item{country}{A character string specifying the country code (e.g., "AT"). This code is used to filter the data files.}

\item{output_sel}{A character string specifying the path to the directory where the generated PDF file should be saved. Defaults to \code{here("output", "plots")}.}

\item{time_min}{A numeric or character string specifying the earliest year to include in the plots (e.g., 1995). Defaults to 1995.}

\item{my_theme}{A ggplot2 theme object to use for the plots. Defaults to \code{ggthemes::theme_fivethirtyeight()}.}

\item{my_colours}{Choose the colours for the lines, for example c("darkred","grey60")}
}
\value{
None. The function generates and saves a PDF file containing the plots.
}
\description{
This function reads annual data from two different sources ("T" and "V")
for a specified country, filters and processes the data, and generates time series plots
comparing the two sources for various statistical concepts (STO).
The plots are saved as a PDF file.
}
\examples{
\dontrun{
nfsa_a_plots_T_V(country = "AT", output_sel = here("my_output"))
}

}
