% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_asset_sync.R
\name{nfsa_asset_sync}
\alias{nfsa_asset_sync}
\title{Synchtonise common files in \verb{M:/nas/Rprod/assets} with local folder
Checks if newer versions are available and copied it to the local folder}
\usage{
nfsa_asset_sync(
  source_dir = "M:/nas/Rprod/assets",
  target_dir = here::here("assets"),
  recursive = FALSE
)
}
\arguments{
\item{source_dir}{"M:/nas/Rprod/assets"}

\item{target_dir}{here::here("assets")}

\item{recursive}{FALSE}
}
\value{
copies more recent files in local folder
}
\description{
Synchtonise common files in \verb{M:/nas/Rprod/assets} with local folder
Checks if newer versions are available and copied it to the local folder
}
\examples{
nfsa_asset_sync()
}
