% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_compare_xml.R
\name{nfsa_compare_xml}
\alias{nfsa_compare_xml}
\title{Compare two xml files from the same country}
\usage{
nfsa_compare_xml(
  path_new_file,
  path_prev_file,
  interactive = TRUE,
  folder_sel = "M:/Incoming_SDMX_files"
)
}
\arguments{
\item{path_new_file}{the most recent file}

\item{path_prev_file}{the file to compare with}

\item{interactive}{if the user choose the files via the menu}
}
\value{
a temporary excel file
}
\description{
Compare two xml files from the same country
}
