% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_completeness_aggregation.R
\name{nfsa_completeness_aggregation}
\alias{nfsa_completeness_aggregation}
\title{NFSA Completeness Aggregation}
\usage{
nfsa_completeness_aggregation(
  country,
  file = FALSE,
  output_sel = here::here("output", "completeness")
)
}
\arguments{
\item{country}{A character vector specifying the country(ies) codes (e.g., "BE").}

\item{file}{If a file should be written. FALSE by default which will open a temporary Excel file.}

\item{output_sel}{A character string specifying the output directory for the completeness report.
Defaults to \code{here::here("output", "completeness")}.}
}
\value{
A data frame containing the missing data points. Returns an empty dataframe if data requirements are fulfilled.
}
\description{
This function checks the completeness of NFSA data for a given country for Table 801,
comparing it against predefined data requirements. It identifies missing data points and
exports them to an Excel file.
}
\examples{
\dontrun{
# Check completeness for Belgium (BE) for table T0801 and save the report to the default output directory
missing_data <- nfsa_completeness_aggregation(country = "BE")
}

}
