% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_completeness_regulation_T0801SA.R
\name{nfsa_completeness_regulation_T0801SA}
\alias{nfsa_completeness_regulation_T0801SA}
\title{Check NFSA Completeness Against Regulation for Table 0801SA}
\usage{
nfsa_completeness_regulation_T0801SA(
  country,
  file = FALSE,
  output_sel = here::here("output", "completeness")
)
}
\arguments{
\item{country}{A character vector specifying the countries for which to check data completeness.}

\item{file}{If a file should be written. FALSE by default which will open a temporary Excel file.}

\item{output_sel}{A file path specifying the directory where the output Excel file should be saved. Defaults to "output/completeness" inside the project.}
}
\value{
Invisible. The function's primary side effect is to produce an Excel file when data is missing.
It also prints messages to the console indicating whether the data requirements are fulfilled or
where the output file is saved.
}
\description{
This function checks the completeness of NFSA data against predefined regulatory requirements.
It identifies missing data points for a specified country and table, and generates an Excel file detailing
these gaps if any exist.
}
\examples{
\dontrun{
# Check completeness for Germany save the output in "my_output" directory
nfsa_completeness_regulation_T0801SA(country = "DE",output_sel = "my_output")
}
}
