% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_folders.R
\name{nfsa_folders}
\alias{nfsa_folders}
\title{Create NFSA Folders}
\usage{
nfsa_folders()
}
\value{
None. This function creates directories and does not return any value.
}
\description{
This function creates a directory structure for the NFSA project,
including folders for data, assets, output, and various subfolders within them.
The directory structure is created relative to the current working directory.
}
\details{
The function uses the \code{fs} package to create the following directory structure:
\itemize{
\item{\code{data/}}
\item{\code{data/a/new/}}
\item{\code{data/a/prev/}}
\item{\code{data/q/new/}}
\item{\code{data/q/prev/}}
\item{\code{data/q/new/nsa/}}
\item{\code{data/q/prev/nsa/}}
\item{\code{data/q/new/sca/}}
\item{\code{data/q/prev/sca/}}
\item{\code{assets/}}
\item{\code{output/}}
\item{\code{output/completeness/}}
\item{\code{output/frequency/}}
\item{\code{output/inter_domain/}}
\item{\code{output/logs/}}
\item{\code{output/negatives_zeroes/}}
\item{\code{output/plots/}}
\item{\code{output/revisions/}}
\item{\code{output/seas/}}
\item{\code{output/sequence/}}
}
}
\examples{
\dontrun{
nfsa_folders()
# Creates the NFSA directory structure in the current working directory.
}

}
