% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_internal_consistency_T0800.R
\name{nfsa_internal_consistency_T0800}
\alias{nfsa_internal_consistency_T0800}
\title{Performs internal consistency checks on NFSA data.}
\usage{
nfsa_internal_consistency_T0800(
  dataset,
  output_sel = here::here("output", "internal"),
  threshold = 2,
  rounding = 1
)
}
\arguments{
\item{dataset}{A data frame containing the NFSA data. This data frame
should include columns for \code{sto} (Statistical Object), \code{ref_sector}
(Reference Sector), \code{accounting_entry}, \code{obs_value} (Observed Value),
\code{ref_area} (Reference Area), and \code{time_period}.}

\item{output_sel}{(optional) A character string specifying the directory where
the output Excel file will be saved. Defaults to \code{"output/internal"}
within the project directory (using \code{here::here()}).}

\item{threshold}{(optional) A numeric value specifying the threshold for
identifying discrepancies.  Differences greater than this threshold
(in absolute value) will be flagged. Defaults to 2.}

\item{rounding}{(optional) An integer specifying the number of decimal places
to round the \code{obs_value} and check calculations to.  Defaults to 1.}
}
\value{
None. The function writes the results to an Excel file.
}
\description{
This function takes an NFSA dataset and performs a series of internal
consistency checks based on predefined rules. It identifies discrepancies
in the data based on a specified threshold and outputs the results to an
Excel file.  The checks include uses vs resources, S1 vs sum of sub-sectors,
subitems vs total, and balancing items.
}
\details{
The function performs the following main categories of checks:

\itemize{
\item{\strong{Uses vs Resources (UR)}: Checks for consistency between
uses and resources for various statistical objects and reference
sectors.}
\item{\strong{S1 vs Sum of Sub-sectors (S1SS)}: Verifies that the
value for sector S1 is consistent with the sum of its sub-sectors.}
\item{\strong{Subitems vs Total (SIT)}: Checks if total values are
equal to the sum of their subitems.}
\item{\strong{Balancing Items (BI)}: Performs checks to ensure that
balancing items are consistent with other related variables.}
}

The specific checks performed within each category depend on the
\code{sto} and \code{ref_sector} values present in the dataset.  The function
dynamically adjusts the checks based on the available data.

The output Excel file contains a separate sheet for each consistency
check that identified discrepancies exceeding the specified threshold. Each
sheet includes the relevant columns from the input dataset, as well as
the calculated check value and the difference between expected and
actual values. "NaN" values are used in the output to represent
missing data to avoid errors in Excel.
}
\examples{
\dontrun{
# Example usage:
nfsa_internal_consistency_T0800(my_dataset, output_sel = "my_output_folder", threshold = 5)
}

}
