% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_loaded.R
\name{nfsa_loaded}
\alias{nfsa_loaded}
\title{Check NFSA Loading Logs}
\usage{
nfsa_loaded(
  time_min,
  time_max = lubridate::ymd(lubridate::today()),
  output_sel = here("output", "logs"),
  recursive = FALSE
)
}
\arguments{
\item{time_min}{A date specifying the minimum time for log files to be considered.
Defaults to the beginning of the current day.}

\item{time_max}{A character string representing the maximum date and time for filtering received files, defaults to current date).  Files received after this time will be excluded.}

\item{output_sel}{Path to the directory where the output Excel file will be saved.
Defaults to "output/logs" within the project directory (using \code{here::here()}).}

\item{recursive}{Logical. Should the function search for log files recursively within subdirectories?
Defaults to \code{FALSE}.}
}
\value{
Invisibly returns a tibble containing the processed log data if any files were loaded, otherwise
invisibly returns \code{NULL}. An Excel file containing the processed data is saved in the
\code{output_sel} directory.
}
\description{
This function scans NFSA loading logs, extracts relevant information,
and saves it to an Excel file.
}
\examples{
\dontrun{
# Check logs loaded since yesterday
nfsa_loaded(time_min = Sys.Date() - 1)

# Check logs loaded since a specific date
nfsa_loaded(time_min = "2025-10-01")

# Check logs recursively within subdirectories
nfsa_loaded(time_min = "2020-01-01",recursive = TRUE)
}
}
