% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_local_parquet.R
\name{nfsa_local_parquet}
\alias{nfsa_local_parquet}
\title{Identify and copy NFSA parquet files}
\usage{
nfsa_local_parquet(country, output_folder)
}
\arguments{
\item{country}{A character vector specifying the country code (e.g., "BE").}

\item{output_folder}{A character string specifying the local folder to copy the files to.}
}
\value{
None. The function copies files as a side effect.
}
\description{
This function identifies and copies NFSA  parquet files from a server location to a local output folder.
}
\details{
This function copies NSA, SCA and A files (both new and previous) from a server location to a local folder. The country code is used to filter the files to copy.
The overwrite argument is set to FALSE, and the copy.date argument is set to TRUE in \code{file.copy}.
The global \code{output_folder} variable needs to be defined outside the function
}
\examples{
\dontrun{
# Example usage:
# Assuming output_folder is already defined, like output_folder <- "C:/users/biedmlu/R/QNFSA/Rprod/data/"
nfsa_local_parquet(country = "BE", output_folder = "C:/users/biedmlu/R/QNFSA/Rprod/data/")
}

}
