% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_move_convert_T0800.R
\name{nfsa_move_convert_T0800}
\alias{nfsa_move_convert_T0800}
\title{Convert SDMX data to Parquet format for NASEC_T08000_A data.}
\usage{
nfsa_move_convert_T0800(
  country,
  year,
  input_sel = "M:/nas/Incoming_SDMX_files/",
  output_sel = here("data", "a", "new")
)
}
\arguments{
\item{country}{A character string specifying the country code.}

\item{year}{A character string specifying the year e.g.,(2024_0000 for 2024)}

\item{input_sel}{A character string specifying the input directory where the SDMX files are located.
Defaults to "M:/nas/Incoming_SDMX_files/".}

\item{output_sel}{A character string specifying the output directory where the Parquet files should be saved.
Defaults to "data/a/new" relative to the project root using \code{here::here()}.}
}
\value{
This function primarily operates for side effects (writing Parquet files).  It prints messages
to the console indicating whether a new file was processed, if no file was found, or if no new file exists.
}
\description{
This function reads SDMX files, converts them to Parquet format, and saves them to a specified directory.
}
\details{
The function searches for SDMX files matching the pattern
"NASEC_T08000_A_{country}_{year}" in the input directory. It identifies the file
with the latest version number. If a new file is found (compared to existing files in the
output directory), it reads the SDMX data, cleans the column names, adds a 'received'
timestamp, selects relevant columns, converts the 'obs_value' column to numeric, adds an
'embargo_date' column if missing and then writes the data to a Parquet file in the output
directory.
}
\examples{
nfsa_move_convert_T0800(country= "BE",
year = "2024_0000",
input_sel = "M:/nas/Incoming_SDMX_files/",
output_sel = "C:/users/biedmlu/data/a/new/")

}
