% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_move_convert_T0800_ASA.R
\name{nfsa_move_convert_T0800_ASA10}
\alias{nfsa_move_convert_T0800_ASA10}
\title{Convert and move T0800 files to parquet format from ASA folder.}
\usage{
nfsa_move_convert_T0800_ASA10(
  country,
  input_sel =
    "M:/nas/ASA10/Production/2024/(1) ASA/(1_1) Original transmission/(1_1_1) Original xml files/",
  output_sel = here("data", "a", "new")
)
}
\arguments{
\item{country}{A character string specifying the country code for the T0800 files.}

\item{input_sel}{A character string specifying the path to the directory containing the
original XML files. Defaults to "M:/nas/ASA10/Production/2024/(1) ASA/(1_1) Original transmission/(1_1_1) Original xml files/".}

\item{output_sel}{A character string specifying the path to the directory where the
converted parquet files will be saved. Defaults to here("data","a","new").}
}
\value{
None. This function saves parquet files to the specified output directory.
}
\description{
This function converts T0800 XML files for a specified country to parquet format,
cleans the data, and saves it to a specified output directory.  It handles variations
in the file structure, including the presence or absence of an embargo date.
}
\examples{
\dontrun{
nfsa_move_convert_T0800_ASA10(country = "BE")
}

}
