% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_move_convert_T0801_QSA.R
\name{nfsa_move_convert_T0801_QSA10}
\alias{nfsa_move_convert_T0801_QSA10}
\title{Convert and move T0801 files to parquet format from QSA folder.}
\usage{
nfsa_move_convert_T0801_QSA10(
  country,
  input_sel =
    "M:/nas/QSA10/Production/2025Q2/(1) QSA/(1_1) Original transmission/(1_1_1) Original xml files/",
  output_sel = here("data", "q", "new", "nsa")
)
}
\arguments{
\item{country}{A character string specifying the country code for the T0801 files.}

\item{input_sel}{A character string specifying the path to the directory containing the
original XML files. Defaults to "M:/nas/QSA10/Production/2025Q2/(1) QSA/(1_1) Original transmission/(1_1_1) Original xml files/".}

\item{output_sel}{A character string specifying the path to the directory where the
converted parquet files will be saved. Defaults to here("data","a","new").}
}
\value{
None. This function saves parquet files to the specified output directory.
}
\description{
This function converts T0801 XML files for a specified country to parquet format,
cleans the data, and saves it to a specified output directory.  It handles variations
in the file structure, including the presence or absence of an embargo date.
}
\examples{
\dontrun{
nfsa_move_convert_T0801_QSA10(country = "BE")
}

}
