% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_obs_status.R
\name{nfsa_obs_status}
\alias{nfsa_obs_status}
\title{Analyze and Report Observation Status}
\usage{
nfsa_obs_status(
  table = "T0801",
  type = "new",
  input_sel = "M:/nas/Rprod/data/",
  output_sel = here::here("output", "flags")
)
}
\arguments{
\item{table}{Character string indicating the table to analyze.  Must be
"T0800" or "T0801".  Defaults to "T0801".}

\item{type}{Character string indicating whether to process "new" or "prev"
(previous) data. Defaults to "new".}

\item{input_sel}{Character string specifying the base path to the data
directory. Defaults to \verb{"M:/nas/Rprod/data/",}.}

\item{output_sel}{Character string specifying the path to the output
directory where the Excel file will be saved. Defaults to
\code{here::here()}.}
}
\value{
None. The function's primary effect is to write an Excel file
containing the analysis results to the specified \code{output_sel}.  It also
prints a success message to the console using \code{cli::cli_alert_success()}.
}
\description{
This function analyzes observation status data from NFSA datasets,
aggregates it by country, statistical object (STO), and time period, and
writes the results to an Excel file. It supports both annual (T0800) and
quarterly (T0801) datasets, as well as new and previous versions.
}
\details{
The function reads data from arrow files located in subdirectories
of the \code{input_sel} directory, based on the \code{table} and \code{type}
parameters. It then joins the data with a lookup table (\code{nfsa::nfsa_sto_lookup}),
performs aggregations, and writes the results to an Excel file. The file
name includes a timestamp and indicates the dataset being analyzed.
}
\examples{
\dontrun{
# Analyze the new T0801 data and save the output to the default location
nfsa_obs_status(table = "T0801", type = "new")

# Analyze the previous T0800 data and save the output to a specific folder
nfsa_obs_status(table = "T0800", type = "prev", output_sel = here::here("my_output"))
}
}
