% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_palettes.R
\name{nfsa_palettes}
\alias{nfsa_palettes}
\title{Function to create palettes from colours}
\usage{
nfsa_palettes(
  name,
  n,
  all_palettes = nfsa_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{name of the palette ("full", "primary", "monoblue", "monofuchsia")}

\item{n}{number of colours}

\item{all_palettes}{list of all palettes}

\item{type}{continue or discrete}
}
\value{
a ready to use palette
}
\description{
Function to create palettes from colours
}
\examples{
scales::show_col(nfsa_palettes(name = "monofuchsia"))
}
