% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_percapita.R
\name{nfsa_percapita}
\alias{nfsa_percapita}
\title{Calculate data per capita}
\usage{
nfsa_percapita(data, freq = "Q")
}
\arguments{
\item{data}{A data frame containing data in millions of monetary units, with columns \code{ref_area} (country code) and \code{time_period} (time period).}

\item{freq}{A character string indicating the frequency of the data ("Q" for quarterly, "A" for annual). Defaults to "Q".}
}
\value{
A data frame with an additional \code{obs_value} column, where the original values have been converted per capita.
}
\description{
Converts data to per capita using population data from Eurostat.
}
\examples{
\dontrun{
# Example usage with quarterly data
data_pop <- nfsa_percapita(data = your_data, freq = "Q")

# Example usage with annual data
data_pop <- nfsa_percapita(data = your_data, freq = "A")
}
}
