% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_q_plots_T0801_BOP.R
\name{nfsa_q_plots_T0801_BOP}
\alias{nfsa_q_plots_T0801_BOP}
\title{Generate plots comparing NFSA and BOP data  for a given country.}
\usage{
nfsa_q_plots_T0801_BOP(
  country,
  quarter,
  output_sel = here::here("output", "plots"),
  time_min = "2015-Q1",
  my_theme = ggthemes::theme_fivethirtyeight(),
  my_colours = c("#B656BD", "#208486", "#AF155C")
)
}
\arguments{
\item{country}{A character string specifying the country code (e.g., "AT"). This code is used to filter the data files.}

\item{quarter}{The quarter from which BOP data should be taken. NFSA by default takes the lates available data.}

\item{output_sel}{A character string specifying the path to the directory where the generated PDF file should be saved. Defaults to \code{here("output", "plots")}.}

\item{time_min}{A character string specifying the earliest time period to include in the plots, in the format "YYYY-QX" (e.g., "1999-Q1"). Defaults to "1999-Q1".}

\item{my_theme}{A ggplot2 theme object to use for the plots. Defaults to \code{ggthemes::theme_fivethirtyeight()}.}

\item{my_colours}{Choose the colours for the lines, for example c("darkred","grey60")}
}
\value{
None. The function generates and saves a PDF file containing the plots.
}
\description{
This function reads T0801 and BOP data for a specified country, filters and processes the data, and generates time series plots
comparing the two sources for various statistical concepts (STO).
The plots are saved as a PDF file.
}
\examples{
\dontrun{
nfsa_remove_output_files("output/plots")
}
}
