% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_read_excel_template_T0800.R
\name{nfsa_read_excel_template_T0800}
\alias{nfsa_read_excel_template_T0800}
\title{Read excel template where T0800 is estimated}
\usage{
nfsa_read_excel_template_T0800(file)
}
\arguments{
\item{file}{}
}
\value{
a data frame
}
\description{
Read excel template where T0800 is estimated
}
\examples{
\dontrun{
nfsa_read_excel_template_T0800("path/to/EE1_2025_F.xlsx")
}
}
