% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_read_xlsx_csv.R
\name{nfsa_read_xlsx_csv}
\alias{nfsa_read_xlsx_csv}
\title{Reads and cleans NFSA data from a semicolon-separated CSV file.}
\usage{
nfsa_read_xlsx_csv(file)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file.}
}
\value{
A data frame containing the cleaned and selected NFSA data.
The data frame includes the following columns: \code{table_identifier},
\code{freq}, \code{adjustment}, \code{ref_area}, \code{counterpart_area}, \code{ref_sector},
\code{counterpart_sector}, \code{consolidation}, \code{accounting_entry}, \code{sto},
\code{instr_asset}, \code{maturity}, \code{expenditure}, \code{unit_measure},
\code{currency_denom}, \code{valuation}, \code{prices}, \code{transformation},
\code{cust_breakdown}, \code{time_period}, \code{obs_value}, \code{obs_status},
\code{conf_status}.
}
\description{
This function reads a CSV file, normally found in \verb{I:/econ/NFSA/OUTPUT/} cleans the column names, and selects
a specific set of columns relevant to NFSA data.
}
\examples{
\dontrun{
# Read an NFSA CSV file
nfsa_data <- nfsa_read_xlsx_csv("I:/econ/NFSA/OUTPUT/SE1_2025Q2_F.xlsx.csv")

# Display the first few rows of the data
head(nfsa_data)
}
}
