% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_report_transmissions.R
\name{nfsa_report_transmissions}
\alias{nfsa_report_transmissions}
\title{Report transmissions of a table for a period}
\usage{
nfsa_report_transmissions(
  table_sel = "T0801_Q",
  country_sel = nfsa_eu27,
  period_sel,
  path_sel,
  output = "excel"
)
}
\arguments{
\item{table_sel}{the table to select. Options are: "T0801_Q", "T0801SA_Q", "T0800_A"}

\item{country_sel}{country or countries for which to find transmissions. By default EU27.}

\item{period_sel}{the period to check.}

\item{output}{An excel file, or an html file or a chart}
}
\value{
an excel file, an html table or a ggplot
}
\description{
Report transmissions of a table for a period
}
\examples{
nfsa_report_transmissions (table_sel = "T0800_A", period_sel = "2024_0000")
}
