% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_revision_eurobase.R
\name{nfsa_revision_eurobase}
\alias{nfsa_revision_eurobase}
\title{Create a virtual database of data sent to eurobase for further processing}
\usage{
nfsa_revision_eurobase(path_sel, table_sel = "nasq_10_nf_tr")
}
\arguments{
\item{path_sel}{Path to the parquet files containing the data}

\item{table_sel}{eurobase table to query. By default "nasq_10_nf_tr". Other options are:
"nasa_10_nf_tr", "nasq_10_ki", "nasa_10_ki"}
}
\value{
an arrow dataset or a data.frame if we write collect().
}
\description{
We pool a set of parquet files into an arrow dataset that can be queried. Do not forget to add
collect() at the end of your query
}
\examples{

nfsa_revision_eurobase(path_sel = "D:/data/eurobase/parquet") |>
dplyr::filter(ref_area == "ES" & sto == "B1GQ") |>
dplyr::collect()

nfsa_revision_eurobase(path_sel = "D:/data/eurobase/parquet",
                      table_sel = "nasa_10_nf_tr") |>
dplyr::filter(ref_area == "ES" & sto == "B1GQ") |>
dplyr::collect()
}
