% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_revision_T0801.R
\name{nfsa_revision_T0801}
\alias{nfsa_revision_T0801}
\title{Detects and reports revisions in T0801 data between two versions of the database.}
\usage{
nfsa_revision_T0801(
  country,
  type = "vintage",
  abs_threshold = 10,
  rel_threshold = 0,
  output_sel = here("output", "revisions")
)
}
\arguments{
\item{country}{A character vector specifying the countries to analyze (ISO2 codes).}

\item{type}{Either to compare with the last version of the previous quarter ("vintage") or with the previous version of this period ("version")}

\item{abs_threshold}{A numeric value specifying the absolute threshold for revision detection. Revisions with an absolute difference greater than this threshold are flagged. Defaults to 10.}

\item{rel_threshold}{A numeric value specifying the relative threshold (percentage of GDP) for revision detection. Revisions with a relative difference greater than this threshold are flagged. Defaults to 0.}

\item{output_sel}{Path to the directory where the output Excel file containing the revisions will be saved. Defaults to \code{here("output", "revisions")}.}
}
\value{
This function does not return a value. It generates an Excel file containing the identified revisions in the specified output directory.  A success message is printed to the console indicating the file name and location.
}
\description{
This function compares new and previous versions of T0801 data, identifies revisions based on absolute and relative thresholds,
and exports the revisions to an Excel file.  It utilizes \code{arrow} for efficient data handling and \code{openxlsx} for Excel output.
}
\examples{
\dontrun{
# Example usage with default thresholds:
nfsa_revision_T0801(country = c("ES", "IT"))

# Example usage with custom thresholds:
nfsa_revision_T0801(country = c("ES", "IT"), abs_threshold = 20, rel_threshold = 0.5)
}

}
