% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_revision_T0801SA.R
\name{nfsa_revision_T0801SA}
\alias{nfsa_revision_T0801SA}
\title{Identify and report revisions in T0801SA data.}
\usage{
nfsa_revision_T0801SA(
  country,
  abs_threshold = 10,
  rel_threshold = 0,
  output_sel = here("output", "revisions")
)
}
\arguments{
\item{country}{A character vector specifying the countries to analyze (ISO2 codes).}

\item{abs_threshold}{A numeric value specifying the absolute threshold for revisions. Defaults to 10.}

\item{rel_threshold}{A numeric value specifying the relative threshold (as \% of GDP) for revisions. Defaults to 0.}

\item{output_sel}{A character string specifying the path to the directory where the revision report should be saved.}
}
\value{
This function does not return a value, but it saves an Excel file containing the identified revisions to the \code{output_sel} directory.  It also prints a success message to the console indicating the location of the file. If no revisions are found, a corresponding message is printed to the console.
}
\description{
This function compares new and previous versions of T0801SA data, identifies
revisions exceeding specified thresholds, and generates an Excel report.
}
\examples{
\dontrun{
# Example usage:
nfsa_revision_T0801SA(country = c("IE", "ES"),
                      abs_threshold = 5,
                      rel_threshold = 0.1,
                      output_sel = here("output", "revisions"))
}

}
