% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_revision_T0801SA_all.R
\name{nfsa_revision_T0801SA_all}
\alias{nfsa_revision_T0801SA_all}
\title{Identify and extract revisions in NASEC T0801sa data across different versions.}
\usage{
nfsa_revision_T0801SA_all(
  country,
  input_sel = "M:/nas/Rprod/data/q/",
  output_sel = here::here("output", "revisions")
)
}
\arguments{
\item{country}{A character string specifying the country code to filter the data.}

\item{input_sel}{A character string specifying the path to the directory containing the Parquet files.
Defaults to \code{"M:/nas/Rprod/data/q/"}.}

\item{output_sel}{Path to the directory where the output Excel file containing the revisions will be saved. Defaults to \code{here("output", "revisions")}.}
}
\value{
An Excel workbook containing the identified revisions,
with columns for the reference area, ID, time period, reference sector, STO, accounting entry,
and the changes between different versions of the data.
}
\description{
This function identifies and extracts revisions in the NASEC T0801 data for a specified country,
comparing different versions of the data stored as Parquet files. It joins the data with a lookup table,
calculates the changes between versions, and prepares the data for export to Excel.
}
\examples{
\dontrun{
# Example usage: Identify and extract revisions for country code "IT"
# from Parquet files located in the default "data/q" directory.
nfsa_revision_T0801SA_all(country = "IT")

# Example usage: Identify and extract revisions for country code "BE"
# from Parquet files located in a custom directory.
nfsa_revision_T0801SA_all(country = "BE", input_sel = "path/to/my/data")
}

}
