% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_scale_colour.R
\name{nfsa_scale_colour}
\alias{nfsa_scale_colour}
\title{Color scale functions for Eurostat palette}
\usage{
nfsa_scale_colour(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link[ggplot2]{scale_colour_manual}}.}
}
\value{
A \code{\link[ggplot2]{Scale}} object that can be added to a ggplot.
}
\description{
This function generates a color scale using the Eurostat palette.
}
