% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_scale_fill.R
\name{nfsa_scale_fill}
\alias{nfsa_scale_fill}
\title{fill discrete palette for charts}
\usage{
nfsa_scale_fill(name = "full", direction = 1)
}
\arguments{
\item{name}{name of the palette. "full" by default. Other options ("primary","monoblue","monofuchsia")}

\item{direction}{of the scale. 1 by default, any other number to reverse it.}
}
\value{
a discrete fill palette
}
\description{
fill discrete palette for charts
}
\examples{
a fill palette
library(ggplot2)

ggplot(iris,aes(Sepal.Length, fill = Species))+
 geom_bar(position = "dodge")+
 scale_fill_nfsa()
}
