% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_scale_fill_c.R
\name{nfsa_scale_fill_c}
\alias{nfsa_scale_fill_c}
\title{fill continuous palette for charts}
\usage{
nfsa_scale_fill_c(name = "monoblue", direction = 1)
}
\arguments{
\item{name}{name of the palette. "full" by default. Other options ("primary","monoblue","monofuchsia")}

\item{direction}{of the scale. 1 by default, any other number to reverse it.}
}
\value{
a continuous fill palette
}
\description{
fill continuous palette for charts
}
\examples{
a continuous fill palette
library(ggplot2)
ggplot(diamonds,aes(cut, color,fill = table))+
geom_tile() +
nfsa_scale_fill_c(name = "monofuchsia", direction = -1)
}
