% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_separate_id.R
\name{nfsa_separate_id}
\alias{nfsa_separate_id}
\title{Separate ID Column in NFSA Data}
\usage{
nfsa_separate_id(data)
}
\arguments{
\item{data}{A data frame containing the \code{id} column to be separated.}
}
\value{
A data frame with the original \code{id} column removed and three new columns: \code{ref_area}, \code{sto}, and \code{accounting_entry}.
}
\description{
This function separates the \code{id} column in an NFSA data frame into three new columns: \code{ref_area}, \code{sto}, and \code{accounting_entry}. The separation is based on the "." delimiter.
}
\examples{
# Create a sample data frame
df <- data.frame(id = c("REF_SECTOR1.STO1.ENTRY1", "REF_SECTOR2.STO2.ENTRY2"))

# Separate the id column
separated_df <- nfsa_separate_id(df)

# Print the result
print(separated_df)

}
