% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_sto_label.R
\docType{data}
\name{nfsa_sto_label}
\alias{nfsa_sto_label}
\title{nfsa_sto label}
\format{
\subsection{\code{nfsa_sto_label}}{

\describe{
\item{sto}{code for sto}
\item{sto_label}{label for sto}
}
}
}
\usage{
nfsa_sto_label
}
\description{
Labels for sto column  taken from the Matis dictionary
}
\keyword{datasets}
