% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_sto_lookup.R
\docType{data}
\name{nfsa_sto_lookup}
\alias{nfsa_sto_lookup}
\title{nfsa_sto_lookup}
\format{
An object of class \code{data.frame} with 2386 rows and 10 columns.
}
\usage{
nfsa_sto_lookup
}
\description{
Correspondance between SDMX/MATIS codes and 3 columns internal code
}
\keyword{datasets}
