% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_T0800_ANA.R
\name{nfsa_T0800_ANA}
\alias{nfsa_T0800_ANA}
\title{Compares NFSA T0800 data with ANA data for consistency.}
\usage{
nfsa_T0800_ANA(
  country,
  year,
  threshold = 1,
  input_sel = "M:/nas/Rprod/data/a/new/",
  output_sel = here::here("output", "inter_domain")
)
}
\arguments{
\item{country}{A character vector specifying the country code(s) to filter data for.}

\item{year}{A numeric value specifying the year for which to compare the data.}

\item{threshold}{A numeric value specifying the threshold for acceptable differences between NFSA and ANA data.
Defaults to 1.}

\item{input_sel}{A file path to the directory containing NFSA data. Defaults to \code{"M:/nas/Rprod/data/a/new"} relative to the project root.}

\item{output_sel}{A file path to the directory where the output Excel file will be saved. Defaults to \code{"output/inter_domain"} relative to the project root.}
}
\value{
This function does not explicitly return a value. It generates an Excel file containing the comparison results
and saves it to the directory specified by \code{output_sel}. A success message is displayed in the console indicating
the file creation and location.  If no inconsistencies are found, a message indicating full consistency is displayed.
}
\description{
This function retrieves and compares NFSA  T0800 data with ANA (Annual National Accounts) data.
It identifies discrepancies between the two datasets based on a specified threshold and generates an Excel file
containing the comparison results.
}
\details{
The function performs the following steps:
\enumerate{
\item Retrieves NFSA T0800 data for the specified country and year using \code{nfsa_get_data()}.
\item Retrieves relevant ANA data from XML files located in a specific directory structure.
\item Cleans and preprocesses both NFSA and ANA data.
\item Joins the two datasets based on common identifiers (ref_area, ref_sector, sto, accounting_entry, time_period).
\item Calculates the difference between NFSA and ANA values and expresses it as a percentage of GDP.
\item Filters the joined data to identify discrepancies exceeding the specified \code{threshold}.
\item Writes the comparison results to an Excel file using \code{openxlsx::write.xlsx()}.
}
}
\examples{
\dontrun{
# Example usage:
nfsa_T0800_ANA(country = "AT", year = 2024, threshold = 0.5)
}

}
