% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_theme.R
\name{nfsa_theme}
\alias{nfsa_theme}
\title{NFSA Theme for ggplot2}
\usage{
nfsa_theme(base_size = 11, base_family = "sans", grid = TRUE, axis = TRUE)
}
\arguments{
\item{base_size}{Numeric. Base font size for the theme (default is 11).}

\item{base_family}{Character. Base font family for the theme (default is "sans").}

\item{grid}{Logical. Whether to display major grid lines (default is TRUE).}

\item{axis}{Logical. Whether to display axis lines (default is TRUE).}
}
\value{
A ggplot2 theme object.
}
\description{
A custom ggplot2 theme similar to Eurostats
}
\examples{
# Apply the NFSA theme to a ggplot
library(ggplot2)

ggplot(data = mtcars, aes(x = mpg, y = disp)) +
  geom_point() +
  nfsa_theme()

# Customize the theme
ggplot(data = mtcars, aes(x = mpg, y = disp)) +
  geom_point() +
  nfsa_theme(base_size = 12, grid = FALSE, axis = FALSE)

\dontrun{
# Example with different fonts (requires the font to be installed)
ggplot(data = mtcars, aes(x = mpg, y = disp)) +
  geom_point() +
  nfsa_theme(base_family = "serif")
}
}
