% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_to_excel.R
\name{nfsa_to_excel}
\alias{nfsa_to_excel}
\title{Export to excel a data frame}
\usage{
nfsa_to_excel(.data)
}
\arguments{
\item{.data}{the data.frame, data.table to be exported}
}
\value{
an excel file in a temporary folder
}
\description{
For closer inspection of small data frames this functions exports it to excel and opens the file.
}
\examples{

 mtcars \%>\%
 nfsa_to_excel()
}
