% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_to_html.R
\name{nfsa_to_html}
\alias{nfsa_to_html}
\title{Create an HTML interactive table with DT}
\usage{
nfsa_to_html(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
DT table
}
\description{
An interactive HTML table with some fancy options (filter,sort,export to excel) is generated from a data frame.
}
\examples{

show_DT(mtcars)

}
