% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_transmissions.R
\name{nfsa_transmissions}
\alias{nfsa_transmissions}
\title{Generate NFSA Transmission Report}
\usage{
nfsa_transmissions(
  time_min,
  time_max = lubridate::ymd(lubridate::today()),
  input_sel = "M:/nas/incoming_SDMX_files/",
  output_sel = here::here("output", "logs"),
  recursive = FALSE
)
}
\arguments{
\item{time_min}{A character string representing the minimum date and time for filtering received files (e.g., "2025-10-20").  Files received before this time will be excluded.}

\item{time_max}{A character string representing the maximum date and time for filtering received files, defaults to current date).  Files received after this time will be excluded.}

\item{output_sel}{A character string specifying the directory where the output Excel file should be saved. Defaults to \code{"output/logs"}.}

\item{recursive}{A logical value. If \code{TRUE}, the function will recursively search for files in subdirectories within the specified path. Defaults to \code{FALSE}.}
}
\value{
This function does not return a value, instead outputs an Excel file with the transmission report.
}
\description{
This function generates a report of NFSA data transmissions received, filtering by a minimum timestamp and outputting the results to an Excel file.
}
\examples{
\dontrun{
# Generate a transmission report for files received since "2025-11-01" and save to default location
nfsa_transmissions(time_min = "2025-11-01")

# Generate a report, save to alternative location and search in subdirectories
nfsa_transmissions(time_min = "2025-10-25", output_sel = "my_output_folder", recursive = TRUE)
}
}
