% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_validation.R
\name{nfsa_validation}
\alias{nfsa_validation}
\title{Find and Compile Validation Timestamps}
\usage{
nfsa_validation(
  time_min,
  time_max = lubridate::ymd(lubridate::today()),
  output_sel = here::here("output", "logs")
)
}
\arguments{
\item{time_min}{A date object or character string coercible to a date,
specifying the minimum date for validation files to be considered.
Files with timestamps earlier than this date will be ignored.}

\item{output_sel}{A character string specifying the directory where the output
Excel file will be saved. Defaults to \code{here::here("output", "logs")}.}

\item{time_ax}{A date object or character string coercible to a date,
specifying the maximum date for validation files to be considered.
Files with timestamps later than this date will be ignored. Defaults to current date}
}
\value{
This function creates and saves an Excel file. If validation files are found, a new excel document is generated in the specified directory.
The file contains the earliest time of validation per combination of table, frequency, adjustment and reference area.
If no validation files are found within the specified time frame, a warning message is displayed.
}
\description{
This function searches for validation files in a specified directory,
extracts the timestamps, and compiles the earliest timestamp for each table,
frequency, adjustment, and reference area combination.
}
\examples{
\dontrun{
nfsa_validation(time_min = "2023-01-01", output_sel = "path/to/output")
}

}
