% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_comext.R
\name{process_comext_data}
\alias{process_comext_data}
\title{Process and aggregate Comext data}
\usage{
process_comext_data(raw_data, quarter, area, config)
}
\description{
Process and aggregate Comext data
}
\keyword{internal}
