% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_comext.R
\name{nfsa_comext}
\alias{nfsa_comext}
\title{Process and export Eurostat Comext data}
\usage{
nfsa_comext(
  area,
  quarter,
  output_sel,
  input_sel =
    "M:/nas/QSA10/Production/2025Q2/(1) QSA/(1_1) Original transmission/(1_1_1) Original xml files/EA early/Comext/"
)
}
\arguments{
\item{area}{A string indicating the geographical area ("EA20" or "EU27").}

\item{quarter}{A string representing the target quarter in the format "YYYYQ#".}

\item{output_sel}{A string specifying the directory where the output CSV file should be saved.}

\item{input_sel}{A string specifying the directory where the additional files are saved. Defaults to
"M:/nas/QSA10/Production/2025Q2/(1) QSA/(1_1) Original transmission/(1_1_1) Original xml files/EA early/Comext/".
It is important that the area has the right associated file. For example, "EA20_lines to add.csv" for area EA20 and "EU27_lines to add.csv" for area EU27.}
}
\value{
A data frame containing the processed Comext data. The data frame is
also written to a CSV file in the specified output directory.
}
\description{
This function retrieves, processes, and exports Eurostat Comext
data for a specified area (EA20 or EU27) and quarter.
}
\examples{
\dontrun{
# Example usage for EA20 in 2023 Q4
nfsa_comext(area = "EA20", quarter = "2023Q4", output_sel = "path/to/output/")

# Example usage for EU27 in 2023 Q4
nfsa_comext(area = "EU27", quarter = "2023Q4", output_sel = "path/to/output/")
}

}
