% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_compare_xml.R
\name{nfsa_compare_xml}
\alias{nfsa_compare_xml}
\title{Compare the content of different xml files}
\usage{
nfsa_compare_xml(paths_sel)
}
\arguments{
\item{paths_sel}{paths to the files}
}
\value{
a data frame
}
\description{
Compare the content of different xml files
}
\examples{
files_sel <- nfsa_find_xml(table_sel = "T0801_Q",country_sel = "BE",period_sel = "2024_0004")
tmp <- nfsa_compare_xml(files_sel)

}
