% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_folders.R
\name{nfsa_folders}
\alias{nfsa_folders}
\title{Create NFSA Folders}
\usage{
nfsa_folders()
}
\value{
None. This function creates directories and does not return any value.
}
\description{
This function creates a directory structure for the NFSA project,
including folders for data, assets, output, and various subfolders within them.
The directory structure is created relative to the current working directory.
}
\details{
The function uses the \code{fs} package to create the the directory structure:
}
\examples{
\dontrun{
nfsa_folders()
# Creates the NFSA directory structure in the current working directory.
}

}
