% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_read_excel_template_T0801.R
\name{nfsa_read_excel_template_T0801}
\alias{nfsa_read_excel_template_T0801}
\title{Read excel template where T0801 is estimates}
\usage{
nfsa_read_excel_template_T0801(file)
}
\arguments{
\item{file}{}
}
\value{
a data frame
}
\description{
Read excel template where T0801 is estimates
}
\examples{
nfsa_read_excel_template_T0801("C:/Users/biedmlu/OneDrive - European Commission/NFSA/EE1_2025Q3_F.xlsx")
}
