% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_read_xml.R
\name{nfsa_read_xml}
\alias{nfsa_read_xml}
\title{Load SDMX XML to Dataframe}
\usage{
nfsa_read_xml(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the SDMX XML file.}
}
\value{
A dataframe containing the data extracted from the SDMX XML file.
}
\description{
This function reads an SDMX XML file and converts it into a dataframe.
}
\examples{
\dontrun{
# Example usage:
file <- "path/to/your/sdmx_file.xml"
df <- nfsa_read_xml(file)
print(head(df))
}

}
