% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_scale_colour.R
\name{nfsa_scale_colour}
\alias{nfsa_scale_colour}
\title{colour palette for charts}
\usage{
nfsa_scale_colour(name = "full", direction = 1)
}
\arguments{
\item{name}{name of the palette. "full" by default. Other options ("primary","monoblue","monofuchsia") as defined in regacc_palettes()}

\item{direction}{of the scale. 1 by default, any other number to reverse it.}
}
\value{
a colour palette
}
\description{
colour palette for charts
}
\examples{
library(ggplot2)
ggplot(iris,aes(Sepal.Length, Sepal.Width, colour = Species))+
geom_point()+
nfsa_scale_colour()
}
