% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_seas_level1.R
\name{nfsa_seas_level1}
\alias{nfsa_seas_level1}
\title{Perform Level 1 Seasonality Checks}
\usage{
nfsa_seas_level1(
  country,
  time_min = "1999-Q1",
  input_sel = "M:/nas/Rprod/data/",
  output_sel = here::here("output", "seas")
)
}
\arguments{
\item{country}{A character vector specifying the countries to process (e.g., "BE", "IT", "FR").  Defaults to c("BE","IT", "FR", "FI", "PL", "EE", "SK", "SI").}

\item{time_min}{From which quarter the series are analysed, by default "1999-Q1"}

\item{input_sel}{The path to the directory containing the input data files. Defaults to \code{"M:/nas/Rprod/data/"}.}

\item{output_sel}{The path to the directory where the output Excel file will be saved. Defaults to \code{here::here("output", "seas")}.}
}
\value{
A data frame containing the results of the level 1 seasonality checks, including the \code{ref_area} (country), \code{id}, and the \code{level1} result (either "PASS" or a specific "FAIL" or "WARNING" message).  Also saves the data frame as an Excel file in the \code{output_sel} directory.
}
\description{
This function performs level 1 seasonality checks on time series data, comparing seasonally adjusted (SCA) and non-seasonally adjusted (NSA) data.
It identifies series with potential issues like residual seasonality, over-adjustment, or inconsistencies between NSA and SCA data.
}
\examples{
\dontrun{
# Perform level 1 seasonality checks for Belgium and Italy,
# using default input and output directories.
results <- nfsa_seas_level1(country = c("BE", "IT"))

# Perform level 1 seasonality checks for France, specifying custom
# input and output directories.
results <- nfsa_seas_level1(country = "FR",
                             input_sel = "/path/to/input/data",
                             output_sel = "/path/to/output/directory")
}

}
