% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_theme.R
\name{nfsa_theme}
\alias{nfsa_theme}
\title{Default theme options for charts}
\usage{
nfsa_theme(font = "serif")
}
\arguments{
\item{font}{Font to be used by default serif}
}
\value{
a set of theme paremeters
}
\description{
Default theme options for charts
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, cyl)) + geom_point() + nfsa_theme(font = "Open Sans")+ ggtitle("Test")
}
