% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfsa_theme.R
\name{nfsa_theme}
\alias{nfsa_theme}
\title{Default theme options for charts}
\usage{
nfsa_theme()
}
\value{
a set of theme parameters
}
\description{
Default theme options for charts
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) + geom_point() + nfsa_theme()+ ggtitle("Test")
}
